chore: move resolve to /common
