+++ src/components/common/Resolve.js
@@ -1,6 +1,21 @@
 import React from "react";
+
 /*
 ce composant permet de resolver plusieurs promises et notifier l'etat via une "render prop" {status, error, result}
+
+const getPromise = () => Promise.resolve()
+
+const Example = () => <Resolve
+  promises={[() => getPromise()]}
+  render={({ status, result }) => (
+    <React.Fragment>
+      {status === "success" && <div>OK: {JSON.stringify(success, null, 2)}</div>}
+      {status === "error" && <div>Impossible de charger la carte des mesures</div>}
+      {status === "loading" && <div>Chargement de la carte des mesures...</div>}
+    </React.Fragment>
+  )}
+/>
+
 */
 class Resolve extends React.Component {
   state = {
